fix(server): remove babel-node
