+++ package.json
@@ -44,7 +44,7 @@
     "minify": "uglifyjs --compress -o dist/bundle.js dist/bundle.js",
     "prod": "npm run build && npm run minify && npm run minify_deps",
     "deploy": "scp -r dist/ 178.170.115.158:tz/frontend/",
-    "start": "npm run build && cd server && babel-node index.js",
+    "start": "npm run build && cd server && node index.js",
     "test": "mocha --compilers js:babel-core/register --recursive src/components/*/*.spec.js",
     "test:watch": "mocha --compilers js:babel-core/register -w --recursive src/components/*/*.spec.js"
   },
